Draw the outlines of shapes when moving objects, eases the correct placement.
git-svn-id: svn://scribus.net/trunk/Scribus@11300 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/canvasmode.cpp
@@ -107,12 +107,12 @@ void CanvasMode::drawSelection(QPainter* psx)
 		{
 			currItem = m_doc->m_Selection->itemAt(a);
 			psx->save();
-			psx->translate(static_cast<int>(currItem->xPos()), static_cast<int>(currItem->yPos()));
+			psx->translate(currItem->xPos(), currItem->yPos());
 			if (currItem->rotation() != 0)
 			{
 				psx->setRenderHint(QPainter::Antialiasing);
 				psx->rotate(currItem->rotation());
-			}		
+			}
 			currItem->paintObj(psx);
 			psx->restore();
 		}
